try to fix topology runs

+++ testsuites/CBLTester/topology_specific_tests/sync_gateway_proxy_enabled/conftest.py
@@ -665,6 +665,23 @@ def setup_customized_teardown_test(params_from_base_test_setup):
         "cbl_db3": cbl_db3,
     }
     log_info("Tearing down test")
-    db.deleteDB(cbl_db1)
-    db.deleteDB(cbl_db2)
-    db.deleteDB(cbl_db3)
+    path = db.getPath(cbl_db1).rstrip("/\\")
+    path2 = db.getPath(cbl_db2).rstrip("/\\")
+    path3 = db.getPath(cbl_db3).rstrip("/\\")
+    if '\\' in path:
+        path = '\\'.join(path.split('\\')[:-1])
+        path2 = '\\'.join(path2.split('\\')[:-1])
+        path3 = '\\'.join(path3.split('\\')[:-1])
+    else:
+        path = '/'.join(path.split('/')[:-1])
+        path2 = '/'.join(path2.split('/')[:-1])
+        path3 = '/'.join(path3.split('/')[:-1])
+    try:
+        if db.exists(cbl_db_name1, path):
+            db.deleteDB(cbl_db1)
+        if db.exists(cbl_db_name2, path2):
+            db.deleteDB(cbl_db2)
+        if db.exists(cbl_db_name3, path3):
+            db.deleteDB(cbl_db3)
+    except Exception as err:
+        log_info("Exception occurred: {}".format(err))
